Config file
A config file has now been added
While still unimplemented, it has Mod Menu integration.
Related to this is the addition of a lang file, and migration from the "server" environment to "*".

And lastly, besides some slight changes to other files, the icon has been changed to a WIP icon for the mod. Cheers!